Add base class for plots
